requirements.txt
streamlit
google_images_search
pillow
requests
python-dotenv